Refactored UI for user details screen and fixed weight date issue
Refactored UI in the UserDetails page to enhance user experience and simplified the usage of date in WeightService. Previously, we were storing and processing the date as an UTC date which resulted in inconsistencies in certain time zones. After the refactor, we now store the date as provided by the user, improving consistency and clarity for the users.