Update WeightRepository and improve UI experience
The WeightRepository was updated to handle conflicts for weight entries by updating weight and date if one for the specific user and date already exists. A significant improvement was made in the UI/UX of the application. The layout of the edit-weight component was adapted to be more responsive and user-friendly by changing the alignment of elements and altering the sizing of containers and buttons.