Update target date logic and improve testing
Updated StatisticsService to use the newestWeightInput date for target if the user target date is null, providing more accurate statistics.
Also, improved the StatisticsTests by reducing input size and adding another method for a separate case of testing current trend without target.
The new test case was added to cover the scenario when users do not set a target date.
Finally, the service singleton of StatisticsService was included in 'Program.cs'.